move model to core folder
