+++ lib/example/example.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:todo_app_class/model/task.dart';
 
 import '../model.dart';
