+++ lib/screens/task_screen.dart
@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:todo_app_class/widgets/tasks_list.dart';
 
-import '../model/TaskData.dart';
-import '../model/task.dart';
+import '../core/model/TaskData.dart';
+import '../core/model/task.dart';
+
 
 class TasksScreen extends StatelessWidget {
   @override
